[FEATURE] #4 - 블록 이동 이벤트, ViewModel 로 이동
